test_switch: deal with sometimes-escaping of []
In some environments, the square brackets of our path filters get
escaped; in some they don't. Not sure exactly what triggers it; it
happens e.g., in Python 2.7 and Python 3.5 on Linux, but not on Python
3.7 on Mac.

It seems we should maybe always escape the URL, but not sure yet.
For now, accept both.