BSC-14299 Timeout Rest Operations
When connecting an optional timeout parameter may be passed,
which is used for the session.

The session timeout value may also be set at any time.

REST calls can take an optional timeout parameter.  If no
timeout parameter is passed the session timeout is used.

Default behaviour is to not time out and wait forever.